Re-introduce AssetPlaybackManager and RemoteCommandManager
Inject apm and rcm into appcoordinator via assetplaybackservice so we have are listening for remote events as early as possible. TODO: on user tap media, get the model object converted to an Asset and set it on the AssetPlaybackManager somehow. inject assetplaybackservice into DemoMusicViewModel at init time? So, when the user taps an asset, the DemoMusicPlayerController could set the asset on AssetPlaybackManager?