disable UserLoginUser-based login(just make fake UserAppUser). hopefully someday we can login for real. Still store a fake password in keychain
Add deleteplaylists api, inject languageservice into playlistsviewmodel. Add swappedlanguage which gets set if the user changes the language during session. The playlistviewmodel detects change, deletes playlists of the visible channel and refetches a playlist in the new language. BUG: the controller is not refreshing on-the-fly, possibly because the view controller does not observe a change so possible TODO: is to add a ValueObserver(?) to refresh the controller UI.